Utit.inherits removed because of is discouraged
In official node.js docs there is written:

Usage of util.inherits() is discouraged. Please use the ES6 class and extends keywords to get language level inheritance support. Also note that the two styles are semantically incompatible.

In this commit this problem is solved. Class syntax is used and replacing prototypes.

Sources for further reading:

> https://nodejs.org/api/util.html#util_util_inherits_constructor_superconstructor
> https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes
> https://github.com/nodejs/node/issues/4179
> https://medium.com/@parsyval/javascript-prototype-vs-class-a7015d5473b