Defined sets of events of imap and notifier
For the purpose of clarify what exactly happens in code, which events are possible, which impossible, and what it does mean there was created list of events.

It is recommended to use these constants instead of names of event in stings because it allow to differentiated between events from different sets.

For example event 'error' or 'end' can be understood differently in context of imap or notifier.