Imap and Notifier events replaced by constants
Instead of simple string in code constants wit events names are applied. It allow to differentiated between events of different objects that can have the same name. It allows to manage names of events independent to logic of code.

Sources:
Pros of this approach (in php but it desn't matter)
> https://pehapkari.cz/blog/2017/07/12/the-bulletproof-event-naming-for-symfony-event-dispatcher/
Example of merge in big js project where constants were used as names of events
> https://github.com/mochajs/mocha/pull/3655

Opposed opinion (why constant is bad idea):
> https://stackoverflow.com/questions/42713927/why-arent-constants-used-for-events-in-node-js

So I invite to the discussion if you think that these changes are bad.

Why `require('events')` instead `require('events').EventEmitter`? Both forms works but proposed in this commit is now recommended and agreed in current documentation. 

Further reading:
Edit of first answer there
> https://stackoverflow.com/questions/21557369/javascript-requireevents-eventemitter
Official docs:
> https://nodejs.org/api/events.html
Code with backward compatibility
> https://github.com/nodejs/node-v0.x-archive/blob/master/lib/events.js